requirements.txt
# Framework principal
streamlit>=1.40.0
crewai>=0.41.0
langchain>=0.3.0
langchain-openai>=0.1.0
langchain-community>=0.3.0

# Base de données et vectorisation
psycopg2-binary==2.9.9
pgvector==0.2.4

# APIs et scraping
openai>=1.14.0
arxiv==2.1.0
requests==2.31.0

# Génération PDF
reportlab==4.1.0
pypdf2==3.0.1

# Utilitaires
python-dotenv==1.0.0
numpy>=1.26.0
tiktoken>=0.6.0

# MCP (Model Context Protocol)
mcp>=0.9.0
nest-asyncio>=1.5.8

# Backend API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

